replace bitset with the github one

+++ lalr1-core/src/lalr1_by_lr0.rs
@@ -11,14 +11,15 @@ use smallvec::SmallVec;
 // inner version, the return value doesn't contain `link`
 fn _lalr1_only<'a>(lr0: &'a Vec<(Vec<LRItem<'a>>, HashMap<u32, u32>)>, g: &'a impl AbstractGrammarExt<'a>) -> Vec<LRState<'a>> {
   let mut ctx = LRCtx::new(g);
-  let mut look_ahead = lr0.iter().map(|(items, _)| vec![BitSet::new(ctx.token_num); items.len()]).collect::<Vec<_>>();
+  let mut look_ahead = lr0.iter()
+    .map(|(items, _)| vec![BitSet::new(ctx.token_num as usize); items.len()]).collect::<Vec<_>>();
   let mut clo_cache = HashMap::new();
   let mut prop = Vec::new();
   let start_prod = g.start().0.as_ref();
 
   for (i, item) in lr0[0].0.iter().enumerate() {
     if item.prod == start_prod {
-      look_ahead[0][i].set(g.eof());
+      look_ahead[0][i].set(g.eof() as usize);
       break;
     }
   }
@@ -30,8 +31,8 @@ fn _lalr1_only<'a>(lr0: &'a Vec<(Vec<LRItem<'a>>, HashMap<u32, u32>)>, g: &'a im
         // ctx.closure is really slow, so add a cache here
         let clo = clo_cache.entry(item.unique_id()).or_insert_with(||
           ctx.closure({
-                        let mut look_ahead = BitSet::new(ctx.token_num);
-                        look_ahead.set(ctx.token_num - 1);
+                        let mut look_ahead = BitSet::new(ctx.token_num as usize);
+                        look_ahead.set(ctx.token_num as usize - 1);
                         let mut init = HashMap::new();
                         init.insert(item, look_ahead);
                         init
@@ -46,7 +47,7 @@ fn _lalr1_only<'a>(lr0: &'a Vec<(Vec<LRItem<'a>>, HashMap<u32, u32>)>, g: &'a im
           let goto_item_idx = lr0[goto_state as usize].0.iter().enumerate().find(|item| item.1.unique_id() == goto_item_id).unwrap().0;
           let goto_look_ahead = &mut look_ahead[goto_state as usize][goto_item_idx];
           goto_look_ahead.or(&clo_item_look_ahead);
-          if clo_item_look_ahead.test(ctx.token_num - 1) {
+          if clo_item_look_ahead.test(ctx.token_num  as usize - 1) {
             prop.push((from, goto_look_ahead.as_mut_ptr()));
           }
         }
@@ -88,11 +89,11 @@ pub fn work<'a>(lr0: &'a Vec<(Vec<LRItem<'a>>, HashMap<u32, u32>)>, g: &'a impl 
     }
     for (item, (_, look_ahead)) in state.iter().zip(result[i].items.iter()) {
       if item.dot == item.prod.len() as u32 {
-        if look_ahead.test(g.eof()) && item.prod_id == start_id {
+        if look_ahead.test(g.eof()  as usize) && item.prod_id == start_id {
           act.insert(eof, smallvec![ParserAct::Acc]);
         } else {
           for i in 0..token_num {
-            if look_ahead.test(i) {
+            if look_ahead.test(i  as usize) {
               // maybe conflict here
               act.entry(i).or_insert_with(|| SmallVec::new()).push(ParserAct::Reduce(item.prod_id));
             }
